requirements.txt
# ── Web Framework ────────────────────────────────────────────
flask>=3.0.0

# ── LangChain Core ───────────────────────────────────────────
# langchain 1.2+ moved chains into langchain-classic
langchain>=1.2.0
langchain-classic>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-pinecone>=0.2.0

# ── Vector Store ─────────────────────────────────────────────
pinecone>=7.0.0

# ── LLM / Embeddings ─────────────────────────────────────────
openai>=2.0.0
sentence-transformers>=2.7.0

# ── Document Loaders ─────────────────────────────────────────
pypdf>=4.0.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
tiktoken>=0.7.0
